Fixed unit tests in src/test/java com.parkit.parkingsystems
Swicthed to callendar as a class as it was already present in the program albeit not used. And Eclipse as well as Oracle indicate, the Date class as being deprecated. Converted what needs to be for UT but TicketDAO needs to be checked out for integration level tests.(To do)